+++ src/main/java/com/eecrs/eecrsbe/service/ZoneService.java
@@ -7,6 +7,6 @@ import java.util.List;
 
 public interface ZoneService {
 
-    Zone findById(long id);
+    Zone findById(Integer id);
     List<Zone> findByRegion(Region region);
 }
